fix: remove tree-sitter from source
