+++ crates/core/src/source.rs
@@ -14,7 +14,6 @@
 use crate::{language::Language, node::KindId, Position};
 use std::borrow::Cow;
 use std::ops::Range;
-use tree_sitter::{InputEdit, Point};
 
 // https://github.com/tree-sitter/tree-sitter/blob/e4e5ffe517ca2c668689b24cb17c51b8c6db0790/cli/src/parse.rs
 #[derive(Debug)]
@@ -22,20 +21,6 @@ pub struct Edit<S: Content> {
   pub position: usize,
   pub deleted_length: usize,
   pub inserted_text: Vec<S::Underlying>,
-}
-
-fn position_for_offset(input: &[u8], offset: usize) -> Point {
-  debug_assert!(offset <= input.len());
-  let (mut row, mut col) = (0, 0);
-  for c in &input[0..offset] {
-    if *c as char == '\n' {
-      row += 1;
-      col = 0;
-    } else {
-      col += 1;
-    }
-  }
-  Point::new(row, col)
 }
 
 /// NOTE: Some method names are the same as tree-sitter's methods.
@@ -84,7 +69,6 @@ pub trait Doc: Clone + 'static {
 pub trait Content: Sized {
   type Underlying: Clone + PartialEq;
   fn get_range(&self, range: Range<usize>) -> &[Self::Underlying];
-  fn accept_edit(&mut self, edit: &Edit<Self>) -> InputEdit;
   /// Used for string replacement. We need this for
   /// indentation and deindentation.
   fn decode_str(src: &str) -> Cow<[Self::Underlying]>;
@@ -99,24 +83,6 @@ impl Content for String {
   type Underlying = u8;
   fn get_range(&self, range: Range<usize>) -> &[Self::Underlying] {
     &self.as_bytes()[range]
-  }
-  fn accept_edit(&mut self, edit: &Edit<Self>) -> InputEdit {
-    let start_byte = edit.position;
-    let old_end_byte = edit.position + edit.deleted_length;
-    let new_end_byte = edit.position + edit.inserted_text.len();
-    let input = unsafe { self.as_mut_vec() };
-    let start_position = position_for_offset(input, start_byte);
-    let old_end_position = position_for_offset(input, old_end_byte);
-    input.splice(start_byte..old_end_byte, edit.inserted_text.clone());
-    let new_end_position = position_for_offset(input, new_end_byte);
-    InputEdit::new(
-      start_byte as u32,
-      old_end_byte as u32,
-      new_end_byte as u32,
-      &start_position,
-      &old_end_position,
-      &new_end_position,
-    )
   }
   fn decode_str(src: &str) -> Cow<[Self::Underlying]> {
     Cow::Borrowed(src.as_bytes())
